# api/openapi.yaml
openapi: 3.0.3
info:
  title: identity
  version: "1"
  description: |
    Who somebody is, once, for every service operated together. Sign in with
    GitHub or Discord, get one account, carry it everywhere.

    This file is the contract. `internal/api` implements it and a test holds
    the two equal: every route is described here, every path here is a route,
    and every response the test suite provokes is validated against these
    schemas. The Go client in this package is generated from it; a TypeScript
    client would be generated from the same file.

    Two kinds of credential exist and the difference is the security model.
    An **account token** comes from signing in here (or from `POST /v1/tokens`)
    and may manage the account. An **application token** comes from
    `POST /v1/exchange` and is bound to one application's origin (its
    `token.audience`); it may call `/v1/whoami` and `/v1/avatar`, revoke
    itself, sign its browser out with `/v1/signout`, and nothing else. A
    browser-facing service must hold only its own application tokens,
    server-side, and must check `token.audience` against its own origin
    before trusting a `whoami` answer.

    An application token minted for a browser that signed in here is linked
    to that sign-in: revoking the sign-in revokes the token, and signing in to
    the same application again from the same sign-in replaces it.
servers:
  - url: /
...
security:
  - bearer: []
paths:
...
  /v1/whoami:
    get:
      operationId: whoami
      tags: [account]
      summary: Who a token belongs to.
      description: |
        The whole integration seam. A consuming service forwards the bearer
        token it holds and gets the account, its provider identities, its
        groups, and a description of the token itself. Check `token.audience`.
      responses:
        "200":
          description: The account behind the token.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Whoami" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/Failure" }
  /v1/avatar:
    get:
      operationId: avatar
      tags: [account]
      summary: The caller's profile photo, at a size.
      parameters:
        - name: size
          in: query
          description: Square display size in pixels, 16 to 1024. Default 128.
          schema: { type: integer, minimum: 16, maximum: 1024 }
        - name: source
          in: query
          description: A specific picture (`upload`, or a provider name) for a picker preview.
          schema: { type: string }
      responses:
        "200":
          description: The image, private and uncacheable.
          content:
            image/*:
              schema: { type: string, format: binary }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "502": { $ref: "#/components/responses/ProviderDown" }
    post:
      operationId: uploadAvatar
      tags: [account]
      summary: Upload a profile photo.
      description: Raw JPEG, PNG or WebP bytes with a matching Content-Type. Selects `upload` as the picture.
      requestBody:
        required: true
        content:
          image/jpeg: { schema: { type: string, format: binary } }
          image/png: { schema: { type: string, format: binary } }
          image/webp: { schema: { type: string, format: binary } }
      responses:
        "201":
          description: Stored.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Avatar" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "413": { $ref: "#/components/responses/TooLarge" }
        "415": { $ref: "#/components/responses/Unsupported" }
        "429": { $ref: "#/components/responses/Throttled" }
        "500": { $ref: "#/components/responses/Failure" }
        "502": { $ref: "#/components/responses/ProviderDown" }
        "503": { $ref: "#/components/responses/Unavailable" }
    put:
      operationId: selectAvatar
      tags: [account]
      summary: Choose which picture is the profile photo.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [source]
              properties:
                source: { type: string, description: "`upload` or a provider name." }
      responses:
        "204": { description: Selected. }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    delete:
      operationId: removeAvatar
      tags: [account]
      summary: Remove the uploaded photo. Provider pictures remain.
      responses:
        "204": { description: Removed. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/Failure" }
...
  /v1/signout:
    post:
      operationId: signOut
      tags: [handoff]
      summary: Sign a browser out here and at every application it signed in to.
      description: |
        Called server-side by an application, with the application token of
        the session it is ending. When that token was minted for a browser
        signed in here, the sign-in is revoked, and with it every application
        token minted from it, this one included. Any other token is revoked
        alone, with anything minted from it. Answers 204 either way.
      responses:
        "204": { description: Signed out. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "500": { $ref: "#/components/responses/Failure" }
  /v1/tokens:
    get:
      operationId: listTokens
      tags: [tokens]
      summary: The caller's live and recently revoked tokens.
      responses:
        "200":
          description: Tokens, newest first.
          content:
            application/json:
              schema:
                type: object
                required: [tokens]
                properties:
                  tokens:
                    type: array
                    items: { $ref: "#/components/schemas/TokenInfo" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "500": { $ref: "#/components/responses/Failure" }
    post:
      operationId: mintToken
      tags: [tokens]
      summary: Mint a named account token.
      description: |
        At most 25 account tokens are live at once; application tokens do not
        count toward it.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
      responses:
        "201":
          description: The token, shown once.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Token" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /v1/tokens/{id}:
    delete:
      operationId: revokeToken
      tags: [tokens]
      summary: Revoke one of the caller's tokens.
      description: |
        Revoking a token revokes every application token minted from it. An
        application token may revoke only itself.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "204": { description: Revoked. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/Failure" }
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      description: "`Authorization: Bearer bsid_...`"
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string, description: "What went wrong, in a sentence." }
    AccountID:
      type: string
      description: A random 12-character hex id. Stable for the life of the account.
      pattern: "^[0-9a-f]{12}$"
      example: 3f9c2a81d04b
    GroupName:
      type: string
      description: Lowercase letters, digits and hyphens; 1 to 63 characters; starts with a letter or digit.
      pattern: "^[a-z0-9][a-z0-9-]{0,62}$"
      example: basically-core
...
    Token:
      type: object
      description: A freshly minted credential. The token itself is shown exactly once.
      required: [token, account, handle]
      properties:
        token: { type: string, description: "`bsid_<id>_<secret>`. Empty when set as a cookie instead." }
        account: { $ref: "#/components/schemas/AccountID" }
        handle: { type: string }
        expires_at: { type: string, format: date-time }
    TokenInfo:
      type: object
      description: A token as described, never including its secret.
      required: [id, name, audience, created_at]
      properties:
        id: { type: string }
        name: { type: string }
        audience:
          type: string
          description: The application origin this token is bound to; empty for an account token.
        created_at: { type: string, format: date-time }
        expires_at: { type: string, format: date-time }
        revoked: { type: boolean }
        current: { type: boolean, description: This is the token making the request. }
...
    Avatar:
      type: object
      required: [id, width, height, source]
      properties:
        id: { type: string, description: Changes whenever the picture does; use it as a cache key. }
        width: { type: integer }
        height: { type: integer }
        source: { type: string, description: "`upload`, `github` or `discord`." }
    Identity:
      type: object
      description: One provider's proof of the account.
      required: [provider, id, handle, proved_at]
      properties:
        provider: { type: string, enum: [github, discord] }
        id: { type: string, description: "The provider's immutable id (GitHub's number, Discord's snowflake)." }
        handle: { type: string, description: The provider login at the last proof. }
        proved_at: { type: string, format: date-time }
        avatar: { $ref: "#/components/schemas/Avatar" }
    Whoami:
      type: object
      required: [account, handle, created_at, identities, groups, token, avatar_upload_enabled]
      properties:
        account: { $ref: "#/components/schemas/AccountID" }
        handle: { type: string }
        created_at: { type: string, format: date-time }
        identities:
          type: array
          items: { $ref: "#/components/schemas/Identity" }
        groups:
          type: array
          description: Every group the account is in, sorted. Always present; empty when none.
          items: { $ref: "#/components/schemas/GroupName" }
        token: { $ref: "#/components/schemas/TokenInfo" }
        avatar: { $ref: "#/components/schemas/Avatar" }
        uploaded_avatar: { $ref: "#/components/schemas/Avatar" }
        avatar_upload_enabled: { type: boolean }
...
  responses:
    BadRequest:
      description: The request was malformed; the error says how.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: No live bearer token.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: The token exists but may not do this; the error says why.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: No such thing.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    TooLarge:
      description: The body is over the limit.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unsupported:
      description: Not a supported media type.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Throttled:
      description: Too many attempts; Retry-After says when.
      headers:
        Retry-After:
          schema: { type: string }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Failure:
      description: This service could not do it.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    Unavailable:
      description: Busy or not configured; try again.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    ProviderDown:
      description: GitHub, Discord or the asset store could not be reached.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }